Created BH-1-PS for Public Sharing
This is an alternative to Creative Commons Share-Alike Attribution, but
designed for any work, such as visual or literary art, or even programs
and their source code.

There may be a problem with the § symbol in ASCII. If so, this will be
updated to Unicode WITHOUT CHANGING THE OPUS NUMBER, since the legal
code wouldn't have changed.